+ Add select target dialog * Fill Today's practice info * fill best practice of last 30 practices - improve UI

+++ RunStatistic/app/src/main/java/com/tantnt/android/runstatistic/utils/BindingUtils.kt
@@ -18,7 +18,7 @@ import com.tantnt.android.runstatistic.models.PracticeModel
 fun TextView.setPracticeDurationFormatted(practice: PracticeModel?) {
     practice?.let {
         if (isPracticeRunning || shouldShowPracticeResult) {
-            text = TimeUtils.convertDutationToFormmated(practice.duration.toLong())
+            text = TimeUtils.convertDutationToFormmated(practice.duration)
         } else
             text = "00:00"
     }
